LPS-54550 rename
